requirements.txt
amqp==5.0.5
attrs==20.3.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.5
fastapi==0.63.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
kombu==5.0.2
mock==4.0.3
newsapi-python==0.2.6
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.17
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.2
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5
